Change some method signatures to be cleaner

+++ Constants.h
@@ -8,6 +8,6 @@ const int kAMPEventUploadThreshold = 30;
 const int kAMPEventUploadMaxBatchSize = 100;
 const int kAMPEventMaxCount = 1000;
 const int kAMPEventRemoveBatchSize = 20;
-const long kAMPEventUploadPeriodSeconds = 30; // 30s
+const int kAMPEventUploadPeriodSeconds = 30; // 30s
 const long kAMPMinTimeBetweenSessionsMillis = 15 * 1000; // 15s
 const long kAMPSessionTimeoutMillis = 30 * 60 * 1000; // 30m
